+++ R/load.R
@@ -2,46 +2,30 @@
 #'
 #' Load LAGOSNE data from local system files
 #'
-#' @param version character LAGOSNE database version string
-#' @param format character choice of rds or sqlite
+#' @param version character LAGOSNE database version string. Defaults to \code{\link{lagosne_version}}
 #' @param fpath file.path optionally specify custom location of LAGOSNE rds file
 #' @export
 #' @importFrom rappdirs user_data_dir
-#' @importFrom dplyr src_sqlite
+#' @importFrom memoise memoise
+#'
 #' @examples \dontrun{
-#' dt  <- lagosne_load("1.054.1")
-#' dt2 <- lagosne_load("1.054.2")
-#'
-#'compare_columns <- function(num){
-#'  identical(dt2$epi_nutr[,num], dt$epi_nutr[,num])
-#'}
-#'names(dt2$epi_nutr)[sapply(13:94, compare_columns)]
-#'
+#' dt  <- lagosne_load("1.087.1")
 #' }
-lagosne_load <- function(version, format = "rds", fpath = NA){
+lagosne_load <- memoise::memoise(function(version = NULL,
+                                          fpath = NA){
+  if(is.null(version)){
+    version <- lagosne_version()
+    message(paste0("Loading LAGOSNE version: ", version))
+  }
 
   if(!is.na(fpath)){
-
-    if(format == "sqlite"){
-      dplyr::src_sqlite(fpath)
-    }else{
       readRDS(fpath)
-    }
-
   }else{
-
-    if(format == "sqlite"){
-      sqlite_path <- paste0(lagos_path(), "LAGOS.sqlite3")
-      stop_if_not_exists(sqlite_path)
-      dplyr::src_sqlite(sqlite_path)
-    }else{
-      rds_path <- paste0(lagos_path(), "data_", version, ".rds")
-      stop_if_not_exists(rds_path)
-      readRDS(rds_path)
-    }
-
+    rds_path <- paste0(lagos_path(), "data_", version, ".rds")
+    stop_if_not_exists(rds_path)
+    readRDS(rds_path)
   }
-}
+})
 
 #' Load depth data
 #'