+++ inst/doc/working_with_lagos.R
@@ -1,1 +1,0 @@
-### This is an R script tangled from 'working_with_lagos.html.asis'